refactor: 사전 검색 로직 훅 분리
