+++ frontend/src/pages/Main/index.tsx
@@ -1,4 +1,3 @@
-import type { DictNameSearchResult } from 'types/api/dictionary';
 import { CgEnter } from 'react-icons/cg';
 import { useNavigate } from 'react-router-dom';
 import SearchBox from 'components/SearchBox';
@@ -11,6 +10,7 @@ import {
   StartButton,
   Wrapper,
 } from './Main.style';
+import useDictionaryNavigate from 'hooks/useDictrionaryNavigate';
 import { URL_PATH } from 'constants/index';
 import logo from 'assets/logo.svg';
 
@@ -21,22 +21,7 @@ const Main = () => {
     navigate(URL_PATH.PET_REGISTER_SEARCH);
   };
 
-  const goToDictDetailPage = (plantId: number) => {
-    navigate(`/dict/${plantId}`);
-  };
-
-  const search = (searchName: string, searchResults?: DictNameSearchResult[]) => {
-    if (!searchName || !searchResults) return;
-
-    const samePlant = searchResults.find(({ name }) => name === searchName);
-
-    if (!samePlant) {
-      navigate(`/dict?search=${searchName}`);
-      return;
-    }
-
-    goToDictDetailPage(samePlant.id);
-  };
+  const { goToProperDictPage, goToDictDetailPage } = useDictionaryNavigate();
 
   return (
     <Wrapper>
@@ -50,7 +35,11 @@ const Main = () => {
       <Logo src={logo} alt="logo" />
       <SearchMessage>피움에 등록된 식물을 검색해 보세요!</SearchMessage>
       <SearchBoxArea>
-        <SearchBox onEnter={search} onNextClick={search} onResultClick={goToDictDetailPage} />
+        <SearchBox
+          onEnter={goToProperDictPage}
+          onNextClick={goToProperDictPage}
+          onResultClick={goToDictDetailPage}
+        />
       </SearchBoxArea>
     </Wrapper>
   );
